Redirect user to original URL after authentication
When using the security bundle of the Symfony framework,
an authentication success handler is responsible for determining
the URL the user will be redirected to after they are successfully
authenticated. In the DefaultAuthenticationSuccessHandler provided
by Symfony that the firewall in this bundle is using, the URL points
to the root of the website (“/”), unless overriden by
the “_target_path” parameter of the request.

In order to redirect the user to the URL they originally requested,
a custom authentication success handler has been configured.
The SamlAuthenticationSuccessHandler uses the “RelayState” request
parameter to correctly redirect the user to their original
destination, as per the SAML2 specification. Then the SamlController
had to be modified to correctly populate the “RelayState” parameter
of the AuthNRequest sent to the SAML2 Identity Provider with the URL
the user tried to access.

# Conflicts:
#	Controller/SamlController.php